CI: Remove LADYBIRD_BOT_TOKEN from the Ladybird checkout step
We discovered the issue with failed deployments was due to an update to
the checkout action that did not interact well with the deploy action.
There is a fix in the deploy action already. See:

https://github.com/JamesIves/github-pages-deploy-action/commit/1d137fdcc70752d4669a19f62ceb3ad44a6ebf5c

However, a tagged release for the deploy action is not yet available,
and pulling the action at a specific commit hash does not work.

For now, let's pin the checkout action at a known-good version. We can
update both actions when the deploy action has a tagged release. In the
meantime, this just removes the token with write-permissions from the
Ladybird checkout.